dialog refactor minus storedMessages issue
